+++ php/logout.php
@@ -1,5 +1,0 @@
-<?php
-require "common.php";
-
-$auth->logOut();
-loggy("debug", "User logged out", "logout", "logout");